+++ sakura_core/env/DllSharedData.h
@@ -149,7 +149,7 @@ struct DllSharedData {
 
 	// プラグイン
 	short						plugCmdIcons[MAX_PLUGIN*MAX_PLUG_CMD];	// プラグイン コマンド ICON 番号	// 2010/7/3 Uchi
-	int							maxToolBarButtonNum;					// ツールバーボタン 最大値			// 2010/7/5 Uchi
+	size_t						maxToolBarButtonNum;					// ツールバーボタン 最大値			// 2010/7/5 Uchi
 
 	// -- -- 保存対象 -- -- //
 	// 設定
